+++ src/haddock/modules/analysis/contactmap/__init__.py
@@ -28,7 +28,7 @@ from haddock.modules.analysis.contactmap.contmap import (
     make_contactmap_report,
     topX_models,
 )
-from haddock.modules.analysis.contactmap.helper import (
+from haddock.libs.libutil import (
     get_available_memory,
     get_necessary_memory,
 )
@@ -82,10 +82,14 @@ class HaddockModule(BaseHaddockModule):
         #  - Get how much memory the current host system has
         #  - If the system has less memory than needed, fail graciously
         current_memory = get_available_memory()
-        needed_memory = get_necessary_memory(models)
+        needed_memory = get_necessary_memory(models) * self.params["ncores"]
         if current_memory < needed_memory:
             self.log(
-                msg=f"Not enough memory to execute `contactmap`, needs {needed_memory:.2f}Gb has {current_memory:.2f}Gb",
+                msg=(
+                    f"Not enough memory to execute `contactmap` "
+                    f"(needs {needed_memory:.2f}Gb has {current_memory:.2f}Gb). "
+                    "! Skipping this module !"
+                ),
                 level="warning",
             )
             self.output_models = models
